refactor: extract mobile awareness and fix collpased bug
